// server/src/schemas/objects/BenefitPackageMedicalServiceCategoryCountOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '@prisma/client';

const Schema: z.ZodType<Prisma.BenefitPackageMedicalServiceCategoryCountOrderByAggregateInput> =
  z
    .object({
      id: z.lazy(() => SortOrderSchema).optional(),
      createdAt: z.lazy(() => SortOrderSchema).optional(),
      updatedAt: z.lazy(() => SortOrderSchema).optional(),
      isPublished: z.lazy(() => SortOrderSchema).optional(),
      isSoftDeleted: z.lazy(() => SortOrderSchema).optional(),
      isActive: z.lazy(() => SortOrderSchema).optional(),
      deactivationReason: z.lazy(() => SortOrderSchema).optional(),
      deactivationDate: z.lazy(() => SortOrderSchema).optional(),
      benefitPackageId: z.lazy(() => SortOrderSchema).optional(),
      medicalServiceCategoryId: z.lazy(() => SortOrderSchema).optional(),
    })
    .strict();

export const BenefitPackageMedicalServiceCategoryCountOrderByAggregateInputObjectSchema =
  Schema;